Get rid of a weird that API nobody was using, and move it to Utilities where it belongs
